Finishing touches
Okay, so all the camera rotations, and movements are correct, i bug fixed a little, but i realize the issue. the movement cant be changed in a void fixedupdate becuase its constantly refreshing the movement back to the original controls. thats the only thing stoping me from completing this project